+++ activerecord/lib/active_record/locking/optimistic.rb
@@ -60,6 +60,7 @@ module ActiveRecord
       end
 
       private
+
         def increment_lock
           lock_col = self.class.locking_column
           previous_lock_value = send(lock_col).to_i
@@ -80,7 +81,8 @@ module ActiveRecord
           return 0 if attribute_names.empty?
 
           lock_col = self.class.locking_column
-          previous_lock_value = send(lock_col).to_i
+          previous_lock_value = read_attribute_before_type_cast(lock_col)
+
           increment_lock
 
           attribute_names += [lock_col]
@@ -91,7 +93,7 @@ module ActiveRecord
 
             affected_rows = relation.where(
               self.class.primary_key => id,
-              lock_col => previous_lock_value,
+              lock_col => previous_lock_value
             ).update_all(
               attributes_for_update(attribute_names).map do |name|
                 [name, _read_attribute(name)]
@@ -104,9 +106,9 @@ module ActiveRecord
 
             affected_rows
 
-          # If something went wrong, revert the version.
+          # If something went wrong, revert the locking_column value.
           rescue Exception
-            send(lock_col + "=", previous_lock_value)
+            send(lock_col + "=", previous_lock_value.to_i)
             raise
           end
         end